Change painter API, add subpixel drawing
