--- devotee/src/visual/color.rs
@@ -1,8 +1,0 @@
-pub use devotee_backend::Converter;
-
-/// General color trait for color mixing.
-pub trait Color {
-    /// Mix two colors.
-    /// The `other` is applied on top of `self`.
-    fn mix(self, other: Self) -> Self;
-}
